Fix // ModeNaming
The modes are always described in the devices manuals. So they probably differ from device to device. 
I guess there is nothing wrong with keeping the names as they are now since they are more or less just placeholders for the devices real modes.

Correct me if I am wrong.